Add overlay for selected views
