+++ WordPress/Classes/ViewRelated/Media/SiteMedia/Views/SiteMediaCollectionCell.swift
@@ -8,7 +8,7 @@ final class SiteMediaCollectionCell: UICollectionViewCell, Reusable {
     private let placeholderView = UIView()
     private var durationView: SiteMediaVideoDurationView?
     private var documentInfoView: SiteMediaDocumentInfoView?
-    private var badgeView: SiteMediaCollectionCellBadgeView?
+    private var selectionView: SiteMediaCollectionCellSelectionOverlayView?
 
     private var viewModel: SiteMediaCollectionCellViewModel?
     private var cancellables: [AnyCancellable] = []
@@ -52,7 +52,7 @@ final class SiteMediaCollectionCell: UICollectionViewCell, Reusable {
         imageView.image = nil
         imageView.alpha = 0
         placeholderView.alpha = 1
-        badgeView?.isHidden = true
+        selectionView?.isHidden = true
         durationView?.isHidden = true
         documentInfoView?.isHidden = true
     }
@@ -104,11 +104,11 @@ final class SiteMediaCollectionCell: UICollectionViewCell, Reusable {
 
     private func didUpdateBadge(_ badge: SiteMediaCollectionCellViewModel.BadgeType?) {
         if let badge {
-            let badgeView = getBadgeView()
-            badgeView.isHidden = false
-            badgeView.setBadge(badge)
+            let selectionView = getSelectionView()
+            selectionView.isHidden = false
+            selectionView.setBadge(badge)
         } else {
-            badgeView?.isHidden = true
+            selectionView?.isHidden = true
         }
     }
 
@@ -193,18 +193,15 @@ final class SiteMediaCollectionCell: UICollectionViewCell, Reusable {
         return durationView
     }
 
-    private func getBadgeView() -> SiteMediaCollectionCellBadgeView {
-        if let badgeView {
-            return badgeView
+    private func getSelectionView() -> SiteMediaCollectionCellSelectionOverlayView {
+        if let selectionView {
+            return selectionView
         }
-        let badgeView = SiteMediaCollectionCellBadgeView()
-        contentView.addSubview(badgeView)
-        badgeView.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            badgeView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -4),
-            badgeView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -4)
-        ])
-        self.badgeView = badgeView
-        return badgeView
+        let selectionView = SiteMediaCollectionCellSelectionOverlayView()
+        contentView.addSubview(selectionView)
+        selectionView.translatesAutoresizingMaskIntoConstraints = false
+        pinSubviewToAllEdges(selectionView)
+        self.selectionView = selectionView
+        return selectionView
     }
 }
